New: Added capabilities to pages and links

+++ deletecourse.php
@@ -1,18 +1,28 @@
 <?php
 
 require_once('../../config.php');
-//require_once('deletecourse_form.php')
 
+require_login();
+if (isguestuser()) {
+	print_error('guestsarenotallowed');
+}
 global $DB;
 
-//Get course ID
-$id = $_GET['id'];
+if(isset($_GET['id'])) {
+	$id = $_GET['id'];
 
-//Delete course record
-if (!$DB->delete_records('block_ps_selfstudy_course', ['id' => $id])) {
-    print_error('inserterror', 'block_ps_selfstudy');
-}
-$url = new moodle_url('/blocks/ps_selfstudy/managecourses.php');
-redirect($url);
+	if (has_capability('block/ps_selfstudy:managecourses', $context, $USER->id)) {
+      //Delete course record
+		if (!$DB->delete_records('block_ps_selfstudy_course', ['id' => $id])) {
+			print_error('inserterror', 'block_ps_selfstudy');
+		}
+		$url = new moodle_url('/blocks/ps_selfstudy/managecourses.php');
+		redirect($url);
+	} else {
+		print_error('nopermissiontoviewpage', 'error', '');
+	}
 
-// form didn't validate or this is the first display
+} else {
+	$url = new moodle_url('/blocks/ps_selfstudy/managecourses.php');
+	redirect($url);
+	}
